+++ test/test_helper.rb
@@ -60,10 +60,8 @@ class AppiumLibCoreTest
         platformName: :android,
         automationName: 'uiautomator2',
         app: 'test/functional/app/api.apk',
-        platformVersion: '8.1.0',
+        platformVersion: '6.0',
         deviceName: 'Android Emulator',
-        appPackage: 'io.appium.android.apis',
-        appActivity: 'io.appium.android.apis.ApiDemos',
         someCapability: 'some_capability',
         unicodeKeyboard: true,
         resetKeyboard: true